Added mining capability

+++ WorkerService.thrift
@@ -14,5 +14,5 @@ service WorkerService extends SharedService.SharedService {
 
 	SharedService.WorkerResponse recvTransactions(1: list<SharedService.Transaction> TransactionsList, 2: map<string, SharedService.DataItem> dataItemMap);
 
-	oneway void mineBlock(1: MasterService.Block block, 2: i16 nonce, 3: i16 interval);
+	oneway void mineBlock(1: MasterService.Block block, 2: i64 nonce, 3: i16 interval);
 }
